게시물 삭제시 확인절차

+++ src/main/webapp/jsp/article/list.jsp
@@ -5,8 +5,8 @@
     pageEncoding="UTF-8"%>
 <%
 List<Map<String, Object>> articleRows = (List<Map<String, Object>>)request.getAttribute("articleRows");
-int cPage=(int)request.getAttribute("page");
-int totalPage=(int)request.getAttribute("totalPage");
+int cPage = (int)request.getAttribute("page");
+int totalPage = (int)request.getAttribute("totalPage");
 %>
 <!DOCTYPE html>
 <html>
@@ -17,8 +17,9 @@ int totalPage=(int)request.getAttribute("totalPage");
 <body>
 
 	<h1>게시물 리스트</h1>
+	
 	<div>
-	<a href="write">글쓰기</a>
+		<a href="write">글쓰기</a>
 	</div>
 	
 	<table border="2" bordercolor="green">
@@ -38,28 +39,25 @@ int totalPage=(int)request.getAttribute("totalPage");
 			<td><%= (int)articleRow.get("id") %></td>
 			<td><%= (LocalDateTime)articleRow.get("regDate") %></td>
 			<td><a href="detail?id=<%= (int)articleRow.get("id") %>"><%= (String)articleRow.get("title") %></a></td>
-			<td><a href="doDelete?id=<%= (int)articleRow.get("id") %>">삭제버튼</a></td>
+			<td><a href="doDelete?id=<%= (int)articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?')==false) retrun false;">삭제버튼</a></td>
 		</tr>
 	<% } %>
 	</table>
 	
 	<style type="text/css">
-	.page>a.red {
-	color:red;
-	}
+		.page > a.red {
+			color:red;
+		}
 	</style>
 	
 	<div class="page">
-	<%for(int i=1;i<=totalPage;i++){ %>
-	<a class="<%=cPage ==i ? "red":"" %>" href="list?page=<%=i %>"><%=i %></a>
-	<%} %>
-		</div>
+		<%for(int i = 1; i <= totalPage; i++){ %>
+			<a class="<%= cPage == i ? "red" : "" %>" href="list?page=<%= i %>"><%= i %></a>
+		<%} %>
+	</div>
+	
 </body>
 </html>
-
-
-
-
